Added authorization to SessionCalendarFacade, error alerts in UI

+++ src/main/domain/facades/SessionCalendarFacade.java
@@ -7,9 +7,11 @@ import java.util.List;
 import java.util.Set;
 
 import main.domain.Member;
+import main.domain.MemberType;
 import main.domain.Session;
 import main.domain.SessionCalendar;
 import main.exceptions.InvalidSessionException;
+import main.exceptions.UserNotAuthorizedException;
 import main.services.Util;
 import persistence.GenericDaoJpa;
 import persistence.SessionCalendarDaoJpa;
@@ -67,7 +69,11 @@ public class SessionCalendarFacade implements Facade {
 		return calendar.academicYearProperty().getValue();
 	}
 
-	public void addSessionCalendar(SessionCalendar calendar) {
+	public void addSessionCalendar(SessionCalendar calendar) throws UserNotAuthorizedException {
+		MemberType memberType = loggedInMemberManager.getLoggedInMember().getMemberType();
+		if (memberType != MemberType.ADMIN && memberType != MemberType.HEADADMIN)
+			throw new UserNotAuthorizedException();
+
 		// TODO we also need to check if there are no already existing sessionCalendars
 		// which would overlap with this one
 
@@ -76,15 +82,16 @@ public class SessionCalendarFacade implements Facade {
 		GenericDaoJpa.commitTransaction();
 	}
 
-	public void editSessionCalendar(SessionCalendar calendar, LocalDate startDate, LocalDate endDate) {
+	public void editSessionCalendar(SessionCalendar calendar, LocalDate startDate, LocalDate endDate)
+			throws UserNotAuthorizedException {
+		MemberType memberType = loggedInMemberManager.getLoggedInMember().getMemberType();
+		if (memberType != MemberType.ADMIN && memberType != MemberType.HEADADMIN)
+			throw new UserNotAuthorizedException();
+
 		GenericDaoJpa.startTransaction();
 		calendar.setStartDate(startDate);
 		calendar.setEndDate(endDate);
 		GenericDaoJpa.commitTransaction();
-	}
-
-	public void deleteSessionCalendar() {
-		// TODO deleteSessionCalendar
 	}
 
 	/*
@@ -131,7 +138,11 @@ public class SessionCalendarFacade implements Facade {
 		}
 	}
 
-	public void addSession(Session session) {
+	public void addSession(Session session) throws UserNotAuthorizedException {
+		MemberType memberType = loggedInMemberManager.getLoggedInMember().getMemberType();
+		if (memberType != MemberType.HEADADMIN && memberType != MemberType.ADMIN)
+			throw new UserNotAuthorizedException();
+
 		// add to calendar
 		calendar.addSession(session);
 
@@ -141,7 +152,16 @@ public class SessionCalendarFacade implements Facade {
 		GenericDaoJpa.commitTransaction();
 	}
 
-	public void editSession(Session session, Session newSession) {
+	public void editSession(Session session, Session newSession) throws UserNotAuthorizedException {
+		MemberType memberType = loggedInMemberManager.getLoggedInMember().getMemberType();
+		if (memberType != MemberType.HEADADMIN && memberType != MemberType.ADMIN)
+			throw new UserNotAuthorizedException();
+
+		// user is an admin but session to edit isn't organized by this admin
+		if (memberType == MemberType.ADMIN
+				&& !session.getFullOrganizerName().equals(loggedInMemberManager.getLoggedInMember().getFullName()))
+			throw new UserNotAuthorizedException();
+
 		// delete the old session from the runtime calendar
 		calendar.deleteSession(session);
 
@@ -165,7 +185,16 @@ public class SessionCalendarFacade implements Facade {
 		GenericDaoJpa.commitTransaction();
 	}
 
-	public void deleteSession(Session session) {
+	public void deleteSession(Session session) throws UserNotAuthorizedException {
+		MemberType memberType = loggedInMemberManager.getLoggedInMember().getMemberType();
+		if (memberType != MemberType.HEADADMIN && memberType != MemberType.ADMIN)
+			throw new UserNotAuthorizedException();
+
+		// user is an admin but session to edit isn't organized by this admin
+		if (memberType == MemberType.ADMIN
+				&& !session.getFullOrganizerName().equals(loggedInMemberManager.getLoggedInMember().getFullName()))
+			throw new UserNotAuthorizedException();
+
 		// delete from calendar
 		calendar.deleteSession(session);
